% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_CAGED.R
\name{available_update_CAGED}
\alias{available_update_CAGED}
\title{Dado uma data inicial verifica-se a disponibilidade dos dados e/ou endereço até a data atual}
\usage{
available_update_CAGED(last.m = "01", last.y = "2007")
}
\arguments{
\item{last.m, last.y}{mês e ano para formar a data inicial, ex.: "12/2019".}
}
\value{
Um conjunto de dados contendo as seguintes colunas:

\item{month }{
mês de referência dos dados consultados;
}
\item{year }{
ano de referência dos dados consultados;
}
\item{available }{
Se TRUE, os dados e endereços estão disponíveis e FALSO, caso contrário;
}
\item{date_check }{
data da verificação.
}
}
\description{
Permite avaliar a disponibilidade dos dados e/ou endereço para todos os meses e anos no
período compreendido entre uma data incial até a data atual.
}
