% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caged_functions.R
\name{baixar_dados_caged}
\alias{baixar_dados_caged}
\title{Baixa CAGED}
\usage{
baixar_dados_caged(
  inicio = "2020-01",
  fim = "2024-06",
  diretorio = "dados_caged"
)
}
\arguments{
\item{inicio}{Data de início no formato "yyyy-mm" (padrão "2020-01").}

\item{fim}{Data de fim no formato "yyyy-mm" (padrão "2024-06").}

\item{diretorio}{Diretório onde os arquivos serão salvos (padrão "dados_caged").}
}
\description{
Esta função baixa arquivos do CAGED para um diretório especificado.
}
