% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ftp_status.R
\encoding{UTF-8}
\name{check_ftp_status}
\alias{check_ftp_status}
\title{<descrição}
\usage{
check_ftp_status(ftp_url = "ftp://ftp.mtps.gov.br/pdet/microdados/")
}
\arguments{
\item{ftp_url}{Uma string contendo o URL do servidor FTP a ser verificado. O valor padrão é
"ftp://ftp.mtps.gov.br/pdet/microdados/".}
}
\value{
Nenhum valor é retornado. A função imprime uma mensagem indicando se o FTP está online ou offline.
}
\description{
Esta função verifica se um servidor FTP está online, dado um URL padrão.
}
\examples{
# Verifica o status do FTP padrão
check_ftp_status()

# Verifica o status de um FTP personalizado
check_ftp_status("ftp://example.com")

}
