% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Função a ser chamada ao carregar o pacote}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Nome da biblioteca.}

\item{pkgname}{Nome do pacote.}
}
\description{
Função a ser chamada ao carregar o pacote
}
