% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caged_functions.R
\encoding{UTF-8}
\name{download_old_caged}
\alias{download_old_caged}
\title{Baixar Arquivos Antigos do CAGED}
\usage{
download_old_caged(
  inicio = "2020-01",
  fim = "2024-06",
  diretorio = "dados_old_caged"
)
}
\arguments{
\item{inicio}{Uma string representando a data de início no formato "AAAA-MM". O padrão é `"2020-01"`.}

\item{fim}{Uma string representando a data de fim no formato "AAAA-MM". O padrão é `"2024-06"`.}

\item{diretorio}{Uma string especificando o diretório onde os arquivos de dados serão salvos. O padrão é `"dados_caged"`.}
}
\value{
Esta função não retorna um valor. Ela realiza suas operações pelos efeitos colaterais (baixando arquivos para o sistema de arquivos local).
}
\description{
Baixa os arquivos antigos do CAGED (Cadastro Geral de Empregados e Desempregados) para o intervalo de datas especificado e os salva em um diretório local.
}
\details{
A função cria uma sequência de datas mensais entre `inicio` e `fim`, constrói as URLs correspondentes para cada arquivo de dados mensal e os baixa no diretório especificado. Se o diretório não existir, ele será criado.
}
\examples{
\dontrun{
# Baixar dados do CAGED de janeiro de 2020 a junho de 2024
download_old_caged(inicio = "2020-01", fim = "2024-06", diretorio = "dados_caged")
}

}
