% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_CAGED.R
\name{filter_CAGED}
\alias{filter_CAGED}
\title{Filtrar os dados do CAGED de acordo com o projeto}
\usage{
filter_CAGED(conn, data.ref = "10/2019", n.ref = 12, par.CAGED = list())
}
\arguments{
\item{conn}{Objeto DBIConnection produzido por DBI::dbConnect().}

\item{data.ref}{Data de referência do projeto.}

\item{n.ref}{Números de meses do CAGED a serem filtrados em relação a \code{data_ref}. Usada para filtrar
os dados do CAGED, por padrão o filtro é realizado para um período de 12 meses.}

\item{par.CAGED}{Configurações usadas no projeto específico que implicam em filtros nos dados do CAGED.}
}
\value{
um \code{tbl_sql} com os dados de acordo com os parâmetros selecionados de um dado projeto. 
Aplicar função \code{dplyr::collect(object)} para carregar os dados na memória.
}
\description{
Acessa o banco de dados do CAGED e aplica filtros nos dados de acordo com os parâmetros 
CAGEG (par_CAGED) estabelecidos pela SIMG para um determinado projeto.
}
\examples{
 

# Ex.: SICRO seção f
par_CAGED <- list(admitidos = c("1", "2"),
                   cor = c("1", "2", "4", "6", "8", "9"),
                 qtd_horas_contrat = 44,
                 ind_trab_parcial = c("1", "0"),
                  ind_trab_intermitente = c("1", "0"),
                 cod_uf = c("35", "33", "32", "31", "41",
                             "43", "42", "13", "14", "16", "15", "17", "11", "12", "21", "22",
                             "23", "24", "26", "25", "28", "27", "29", "51", "50", "52", "53"),
                  porte_empresa = c("5", "6", "7", "8", "9"),
                  escolaridade = c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11"),
                   sexo = c("1", "2"))
# Consultando CBO's do respectivo projeto (proj) na tabela DeProjParaCBO:
# par.CAGED$cod_cbo <-  
# conn \%>\% 
# dplyr::tbl(dplyr::sql(query_DeProjParaCBO(list(projeto = proj, data_ref = data.ref)))) \%>\% 
# dplyr::select(cod_cbo) \%>\% 
# dplyr::collect() \%>\% 
# unlist() \%>\% 
# unique() \%>\% 
# na.omit()

# con <- DBI::dbConnect(RSQLite::SQLite(), "/home/rumenick/Documentos/data/MDO.sqlite")
# SICRO_sec_f <- filter_CAGED(data.ref = "10/2019", n.ref = 12, par.CAGED = par_CAGED, conn = con)
# dplyr::collect(SICRO_sec_f) # Carregar dados na memória.


}
\author{
Rumenick Pereira da Silva
}
