% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_CAGED.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Metadados de verificação de atualização dos dados do CAGED}
\format{
A data frame with 161 rows and 5 variables
}
\usage{
data(metadata)
}
\description{
Metadados de verificação de atualização dos dados do CAGED
}
\details{
\itemize{
  \item month Mês do ano dos dados verificados atualização;
  \item year Ano dos dados verificados atualização;
  \item date data da verificação da disponibilidade dos dados ou endereço;
  \item available Se TRUE, os Dados e endereços estão disponíveis e FALSO, caso contrário;
  \item description Status dos dados e/ou endereço.
}
}
\references{
ftp://ftp.mtps.gov.br/pdet/microdados/CAGED
}
\keyword{datasets}
