% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_CAGED.R
\name{read_CAGED}
\alias{read_CAGED}
\title{Importar dados do CAGED}
\usage{
read_CAGED(month = "12", year = "2019")
}
\arguments{
\item{month}{Mês do ano qual deseja os dados, deve ser especificado como caracter com dois dígitos,
ex.: "01" (mês de janeiro).}

\item{year}{Ano qual deseja os dados, deve ser especificado como caracter com quatro dígitos,
ex.: "2019".}
}
\description{
Função para leitura dos dados do Cadastro Geral de Empregados e Desempregados
(CAGED) do governo federal do Brasil.
}
\details{
Instituído pela Lei n° 4.923/65, que obrigou as empresas abrangidas pelo Sistema de
Consolidação das Leis do Trabalho a informarem ao Ministério do Trabalho, em relação nominal, a
movimentação de seus empregados, o CAGED tem periodicidade mensal e suas informações destinam-se
a acompanhar e fiscalizar o processo de admissão e dispensa dos trabalhadores, bem como subsidiar
a adoção de medidas contra o desemprego e o estabelecimento de mecanismos de assistência aos
desempregados.
}
\examples{

## Baixa dados do mês de fevereiro de 2018
# CAGEDEST_022018 <- read_CAGED(month = "02", year = "2018")
# CAGEDEST_022018

## Dois ou mais meses ou anos:
# install.packages("dplyr")

# configs <- expand.grid(meses = c("01", "02"), anos = c("2018", "2019"))

# CAGED_list <- mapply(FUN = function(m, y) {read_CAGED(month = m, year = y)},
#                     m = configs$meses, y = configs$anos)
# dplyr::bind_rows(CAGED_list) # Juntar todos os dados

}
\references{
http://portalfat.mte.gov.br/programas-e-acoes-2/caged-3/
}
\author{
Rumenick Pereira da Silva
}
\keyword{Brasil}
\keyword{CAGED}
\keyword{baixar}
\keyword{download}
\keyword{importar}
