% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caged_functions.R
\encoding{UTF-8}
\name{read_old_caged}
\alias{read_old_caged}
\title{Lê e combina arquivos CAGED antigos em um único data frame}
\usage{
read_old_caged(diretorio = "dados_caged")
}
\arguments{
\item{diretorio}{Uma string que especifica o diretório contendo os arquivos `.7z` do CAGED a serem lidos. O padrão é `"dados_caged"`.}
}
\value{
Um data frame contendo os dados combinados de todos os arquivos `.7z` lidos.
}
\description{
Esta função busca todos os arquivos com extensão `.7z` em um diretório especificado, lê cada um usando a função `read_csv2` do pacote `readr` com a codificação `ISO-8859-2`, ajusta os tipos de colunas conforme necessário e combina todos os data frames resultantes em um único data frame.
}
\details{
A função utiliza `list.files` para listar todos os arquivos com extensão `.7z` no diretório especificado. Para cada arquivo, lê o conteúdo usando `read_csv2` com a codificação `ISO-8859-2`. As colunas são automaticamente convertidas para os tipos apropriados usando `type_convert`. Se houver alguma coluna específica que apresente problemas de leitura, você pode ajustar o parâmetro `col_types` dentro da função `read_csv2`, substituindo `"problematic_column"` pelo nome da coluna problemática e definindo seu tipo apropriado (por exemplo, `col_character`).
}
\examples{
\dontrun{
# Lê os arquivos .7z do diretório padrão "dados_caged"
caged_data <- read_old_caged()

# Lê os arquivos .7z de um diretório específico
caged_data <- read_old_caged(diretorio = "meu_diretorio")
}
}
