% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_CAGED.R
\name{repository_update_CAGED}
\alias{repository_update_CAGED}
\title{Verifica alterações nas pastas de dados do repositório do CAGED}
\usage{
repository_update_CAGED()
}
\value{
Um conjunto de dados contendo as seguintes colunas:

\item{url_path }{
link do diretório que estou verificando alterações no repositório;
}
\item{dir_name }{
nome do diretório (ano de referência dos dados);
}
\item{date_update }{
data da última alteração em pelos um dos arquivos do diretório;
}
\item{date_check }{
data da verificação.
}
}
\description{
Acessa a ftp onde está disponível os dados do CAGED e retorna metadados das pastas de dados 
do reposítorio.
}
\references{
ftp://ftp.mtps.gov.br/pdet/microdados/CAGED
}
