% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caged_functions.R
\name{show_files_ftp}
\alias{show_files_ftp}
\title{Verifica arquivos no FTP do CAGED para um ano específico
Esta função verifica e retorna os arquivos disponíveis para um ano específico no FTP do CAGED.
O usuário precisa informar o ano no formato "YYYY" (exemplo: "2024").}
\usage{
show_files_ftp(ano)
}
\arguments{
\item{ano}{Ano no formato "YYYY" (exemplo: "2024").}
}
\value{
Uma lista com os arquivos encontrados no FTP para o ano especificado.
}
\description{
Verifica arquivos no FTP do CAGED para um ano específico
Esta função verifica e retorna os arquivos disponíveis para um ano específico no FTP do CAGED.
O usuário precisa informar o ano no formato "YYYY" (exemplo: "2024").
}
