% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_caged.R
\name{summary_caged}
\alias{summary_caged}
\title{Resume os dados por UF de acordo com o DeProjParaCBO}
\usage{
summary_caged(conn.caged, conn.auxiliar, proj, data.ref.geral)
}
\arguments{
\item{conn.caged}{Objeto DBIConnection produzido por DBI::dbConnect() do BD MDO.}

\item{conn.auxiliar}{Objeto DBIConnection produzido por DBI::dbConnect() do BD DadosAuxiliares.}

\item{proj}{Nome do projeto (Padronizar nomes no futuro).}

\item{data.ref.geral}{Data de referência do projeto.}

\item{par.caged}{Configurações usadas no projeto específico que implicam em filtros nos dados do caged.}
}
\description{
Permite obter as estatísticas da variável salário de acordo com o DeProjParaCBO do respectivo
projeto.
}
\examples{
 

# system.time(summary_caged(conn.caged = DBI::dbConnect(RSQLite::SQLite(), "/home/rumenick/Documentos/MDO.sqlite"), 
# conn.auxiliar = DBI::dbConnect(RSQLite::SQLite(), "/home/rumenick/Documentos/MDODadosAuxiliares.sqlite"), 
# proj = "SICRO_secao_f", 
# data.ref.geral = "01/2020") -> teste)

}
