% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_CAGED.R
\name{filter_CAGED}
\alias{filter_CAGED}
\title{Filtrar os dados do CAGED de acordo com o projeto}
\usage{
filter_CAGED(
  proj = "SICRO_secao_f",
  data.ref = "10/2019",
  n.ref = 12,
  par.CAGED = list(),
  conn
)
}
\arguments{
\item{proj}{Nome do projeto (Padronizar nomes no futuro).}

\item{data.ref}{Data de referência do projeto.}

\item{n.ref}{Números de meses do CAGED a serem filtrados em relação a \code{data_ref}. Usada para filtrar
os dados do CAGED, por padrão o filtro é realizado para um período de 12 meses.}

\item{par.CAGED}{Configurações usadas no projeto específico que implicam em filtros nos dados do CAGED.}

\item{conn}{Objeto DBIConnection produzido por DBI::dbConnect().}
}
\value{
um \code{tbl_sql} com os dados de acordo com os parâmetros selecionados de um dado projeto. 
Aplicar função \code{dplyr::collect(object)} para carregar os dados na memória.
}
\description{
Acessa o banco de dados do CAGED e aplica filtros nos dados de acordo com os parâmetros 
CAGEG (par_CAGED) estabelecidos pela SIMG para um determinado projeto.
}
\examples{
 


}
\author{
Rumenick Pereira da Silva
}
