% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caged_functions.R
\encoding{UTF-8}
\name{read_caged}
\alias{read_caged}
\title{Lê e combina arquivos CAGED compactados em um único data.frame}
\usage{
read_caged(diretorio = "dados_caged")
}
\arguments{
\item{diretorio}{Caminho para o diretório que contém os arquivos `.7z` do CAGED. O padrão é `"dados_caged"`.}
}
\value{
Um `data.frame` contendo os dados combinados de todos os arquivos `.7z` encontrados no diretório.
}
\description{
Esta função busca todos os arquivos com extensão `.7z` em um diretório especificado, lê cada um usando `readr::read_csv2()`, converte os tipos de dados com `readr::type_convert()` e combina todos os data.frames resultantes em um único data.frame.
}
\details{
Os arquivos `.7z` devem conter dados em formato CSV compatível com a função `readr::read_csv2()`. Certifique-se de que todos os arquivos no diretório tenham a mesma estrutura de dados para evitar erros na combinação.
}
\examples{
\dontrun{
# Lê os arquivos do diretório padrão "dados_caged"
dados_caged <- read_caged()

# Lê os arquivos de um diretório específico
dados_caged <- read_caged(diretorio = "meu_diretorio_caged")
}
}
