% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caged_functions.R
\name{read_caged}
\alias{read_caged}
\title{Junta CAGED}
\usage{
read_caged(diretorio = "dados_caged")
}
\arguments{
\item{diretorio}{Diretório onde os arquivos .7z estão salvos (padrão "dados_caged").}
}
\value{
Um data frame com todos os dados combinados.
}
\description{
Esta função lê e junta arquivos CAGED salvos em um diretório especificado.
}
