% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_CAGED.R
\name{summary_CAGED}
\alias{summary_CAGED}
\title{Resume os dados por UF de acordo com o DeProjParaCBO}
\usage{
summary_CAGED(
  conn,
  proj,
  data.ref.geral,
  data.ref.coleta,
  n.ref.coleta = 12,
  par.CAGED
)
}
\arguments{
\item{conn}{Objeto DBIConnection produzido por DBI::dbConnect().}

\item{proj}{Nome do projeto (Padronizar nomes no futuro).}

\item{data.ref.geral}{Data de referência do projeto.}

\item{data.ref.coleta}{Data de coleta dos dados do CAGED, referencia o mês/ano até onde pegar os dados.}

\item{n.ref.coleta}{Números de meses do CAGED a serem filtrados em relação a \code{data.coleta}. Usada para filtrar
os dados do CAGED, por padrão o filtro é realizado para um período de 12 meses.}

\item{par.CAGED}{Configurações usadas no projeto específico que implicam em filtros nos dados do CAGED.}
}
\description{
Permite obter as estatísticas da variável salário de acordo com o DeProjParaCBO do respectivo
projeto.
}
